% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpFiles_Calc Gamma_fromVAR.r
\name{calc.Gamma.fromVAR}
\alias{calc.Gamma.fromVAR}
\title{Calculates the stationary covariance matrix from the discrete-time (un)standardized lagged effects matrix and residual covariance matrix}
\usage{
calc.Gamma.fromVAR(Phi, SigmaVAR)
}
\arguments{
\item{Phi}{(Un)standardized lagged effects matrix.}

\item{SigmaVAR}{Residual covariance matrix of the first-order discrete-time vector autoregressive (DT-VAR(1)) model.}
}
\value{
Stationary covariance matrix, that is, the contemporaneous covariance matrix of the data.
}
\description{
Calculates the stationary covariance matrix from the discrete-time (un)standardized lagged effects matrix and residual covariance matrix
}
\examples{

q = 2
Phi <- matrix(c(0.50, 0.15, 0.25, 0.40), byrow=T, ncol = q)
SigmaVAR <- diag(q) # for ease
calc.Gamma.fromVAR(Phi, SigmaVAR)
}
