% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpFiles_Calc Max Delta t.r
\name{calc.MaxDeltaT}
\alias{calc.MaxDeltaT}
\title{Time-interval (DeltaT) for which Phi_ij(DeltaT) reaches its minimum or maximum (together with that minimum or maximum)}
\usage{
calc.MaxDeltaT(B = NULL, Phi = NULL)
}
\arguments{
\item{B}{Matrix of size q times q of (un)standardized continuous-time lagged effects, called drift matrix. Note that the input uses B for which Phi(DeltaT) = expm(-B*DeltaT), that is, B = -A (it will detect if A is used and continuous with a message).}

\item{Phi}{Optional. Matrix of size q times q of (un)standardized lagged effects. By default, input for B is used; only when B = NULL, Phi will be used (and the corresponding B will be used).}
}
\value{
The output renders, per element (i,j), the time-interval for which Phi_ij reaches its minimum/maximum together with this minimum/maximum.
}
\description{
Time-interval (DeltaT) for which Phi_ij(DeltaT) reaches its minimum or maximum (together with that minimum or maximum). The interactive web application 'Phi-and-Psi-Plots and Find DeltaT' also contains this functionality, you can find it on my website: \url{https://www.uu.nl/staff/RMKuiper/Websites\%20\%2F\%20Shiny\%20apps}.
}
\examples{

##################################################################################################
# Input needed in examples below with q=2 variables.
# I will use the example matrices stored in the package:
Phi <- myPhi[1:2, 1:2]
##################################################################################################

calc.MaxDeltaT(Phi = Phi)

# If you would use the drift matrix Drift as input (with Phi(DeltaT) = expm(-Drift*DeltaT)):
#calc.MaxDeltaT(B = Drift)
#calc.MaxDeltaT(Drift)


# Note that the function 'PhiPlot' can help to see (per element) whether a minimum or maximum is reached.

}
