% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc Gamma.r
\name{Gamma.fromCTM}
\alias{Gamma.fromCTM}
\title{Calculate the stationary covariance matrix Gamma}
\usage{
Gamma.fromCTM(Drift, Sigma)
}
\arguments{
\item{Drift}{(Un)standardized lagged effects matrix of the first-order continuous-time (CT-VAR(1)) model.
It also takes a fitted object from the class "ctsemFit" (from the ctFit() function in the ctsem package); see example below. From such an object, the Drift and Sigma matrices are extracted.}

\item{Sigma}{Residual covariance matrix of the first-order continuous-time (CT-VAR(1)) model, that is, the diffusion matrix.}
}
\value{
This function returns the stationary covariance matrix, that is, the contemporaneous covariance matrix of the data.
}
\description{
Calculates the stationary covariance matrix (Gamma) from the continuous-time (un)standardized lagged effects matrix (Drift) and residual covariance matrix (Sigma). There is also an interactive web application on my website: Standardizing and/or transforming lagged regression estimates (\url{https://www.uu.nl/staff/RMKuiper/Websites\%20\%2F\%20Shiny\%20apps}).
}
\examples{

# library(CTmeta)

## Example 1 ##
#
Phi <- myPhi[1:2,1:2]
#Phi <- matrix(c(0.50, 0.15, 0.25, 0.40), byrow=T, ncol = 2)
if (!require("expm")) install.packages("expm") # Use expm package for function logm()
library(expm)
DeltaT <- 1
Drift <- logm(Phi)/DeltaT
#
q <- dim(Phi)[1]
Sigma <- diag(q) # for ease
#
Gamma.fromCTM(Drift, Sigma)


## Example 2: input from fitted object of class "ctsemFit" ##
#
#data <- myData
#if (!require("ctsemFit")) install.packages("ctsemFit")
#library(ctsemFit)
#out_CTM <- ctFit(...)
#Gamma.fromCTM(out_CTM)

}
