% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myData.R
\docType{data}
\name{myData}
\alias{myData}
\title{An example of data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 2 columns.
}
\usage{
data(myData)
}
\description{
An example of data; for q = 2 variables and 1000 observations.
}
\examples{
data(myData)

}
\keyword{datasets}
