% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDrift.R
\docType{data}
\name{myDrift}
\alias{myDrift}
\title{An example of Drift}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 2 columns.
}
\usage{
data(myDrift)
}
\description{
An example of Drift: a 2x2 lagged effects matrix; hence, for q = 2 variables
}
\examples{
data(myDrift)

}
\keyword{datasets}
