% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGamma.R
\docType{data}
\name{myGamma}
\alias{myGamma}
\title{An example of Gamma}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 2 columns.
}
\usage{
data(myGamma)
}
\description{
An example of Gamma: a stacked matrix of covariance matrices of the data; for q = 2 variables and S = 3 primary studies.
}
\examples{
data(myGamma)

}
\keyword{datasets}
