% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPhi.R
\docType{data}
\name{myPhi}
\alias{myPhi}
\title{An example of Phi}
\format{An object of class \code{matrix} with 6 rows and 2 columns.}
\usage{
data(myPhi)
}
\description{
An example of Phi: a stacked matrix of lagged effects matrices; for q = 2 variables and S = 3 primary studies.
}
\examples{
data(myPhi)

}
\keyword{datasets}
