% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySigmaVAR.R
\docType{data}
\name{mySigmaVAR}
\alias{mySigmaVAR}
\title{An example of SigmaVAR}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 2 columns.
}
\usage{
data(mySigmaVAR)
}
\description{
An example of SigmaVAR: a stacked matrix of residual covariance matrices; for q = 2 variables and S = 3 primary studies. The dimensions of SigmaVAR are (S*q) x q.
}
\examples{
data(mySigmaVAR)

}
\keyword{datasets}
