% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpFiles_Calc CovMx stand Phi.r
\name{calc.CovMxStandPhi}
\alias{calc.CovMxStandPhi}
\title{Calculates the (vectorized) standardized lagged effects matrix, their covariance matrix, and corresponding elliptical 95 percent confidence interval}
\usage{
calc.CovMxStandPhi(N, Phi, Gamma, SigmaVAR, alpha = 0.05)
}
\arguments{
\item{N}{Number of persons (panel data) or measurement occasions - 1 (time series data). Used in determining the covariance matrix of the vectorized standardized lagged effects.}

\item{Phi}{(Un)standardized lagged effects matrix. If necessary, it is standardized and for the standardized and vectorized Phi the covariance matrix is determined.}

\item{Gamma}{Stationary covariance matrix, that is, the contemporaneous covariance matrix of the data.}

\item{SigmaVAR}{Residual covariance matrix of the first-order discrete-time vector autoregressive (DT-VAR(1)) model}

\item{alpha}{The alpha level in determining the (1-alpha)*100 percent confidence interval (CI). By default, alpha is set to 0.05, resulting in a 95& CI}
}
\value{
vectorized standardized lagged effects, their covariance matrix, and the corresponding elliptical/multivariate 95 percent CI.
}
\description{
Calculates the (vectorized) standardized lagged effects matrix, their covariance matrix, and corresponding elliptical 95 percent confidence interval
}
\examples{

# Example where Phi and SigmaVAR are known
N <- 643
q <- 2
Phi <- matrix(c(0.25, 0.10,
                0.20, 0.36), byrow=T, ncol = q)
SigmaVAR <- diag(q) # for ease
Gamma <- calc.Gamma.fromVAR(Phi, SigmaVAR)
calc.CovMxStandPhi(N, Phi, Gamma, SigmaVAR)

# Example where Phi and Gamma are known
N <- 643
q <- 2
Phi <- matrix(c(0.25, 0.10,
                0.20, 0.36), byrow=T, ncol = q)
# Use Gamma from above
SigmaVAR <- Gamma - Phi \%*\% Gamma \%*\% t(Phi)
calc.CovMxStandPhi(N, Phi, Gamma, SigmaVAR)
}
