% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi-plot.R
\name{PhiPlot}
\alias{PhiPlot}
\title{PhiPlot: visualize how lagged parameters vary as function of the time interval.}
\description{
This function generates a plot illustrating how the lagged parameters vary as a function of the time-interval (DeltaT). The following options are available as input: the lagged effects matrix (Phi) from one primary study, the overall Phi matrix output from CTmeta or the underlying drift matrix (Drift) corresponding to either of the first two options are required as input.  An interactive web application offering the same functionality can be found on my website: Phi-and-Psi-Plots and Find DeltaT (\url{https://www.uu.nl/staff/RMKuiper/Websites\%20\%2F\%20Shiny\%20apps}). 
}
\usage{
PhiPlot(
  DeltaT = 1,
  Phi = NULL,
  Drift = NULL,
  Stand = 0,
  SigmaVAR = NULL,
  Sigma = NULL,
  Gamma = NULL,
  Min = 0,
  Max = 10,
  Step = 0.05,
  WhichElements = NULL,
  Labels = NULL,
  Col = NULL,
  Lty = NULL,
  Title = NULL,
  MaxMinPhi = FALSE
)
}
\arguments{
\item{DeltaT}{Optional. The observed time interval. By default, DeltaT = 1.}

\item{Phi}{Matrix of size q x q of (un)standardized discrete-time lagged effects arising from one primary study.  Alternatively, the overall Phi matrix that is output by CTmeta can be used. Note,  the Phi (or Drift) matrix should be standardized to make a fair comparison between cross-lagged effects.  It also takes a fitted object from the classes "varest" (from the VAR() function in vars package) and "ctsemFit" (from the ctFit() function in the ctsem package); see example below. From such an object, the (standardized) Drift matrix is calculated/extracted.}

\item{Drift}{Optional (either Phi or Drift). Underlying continuous-time lagged effects matrix (i.e., Drift matrix) of the discrete-time lagged effects matrix Phi(DeltaT). By default, input for Phi is used; only when Phi = NULL, will Drift be used.}

\item{Stand}{Optional. Indicator for whether Phi (or Drift) should be standardized (1) or not (0). In case Stand = 1, one of the following matrices should be input as well: SigmaVAR, Sigma, or Gamma  (or it is subtracted from a varest or ctsemFit object). By default, Stand = 0.}

\item{SigmaVAR}{Optional (if Stand = 1, then either SigmaVAR, Sigma, or Gamma needed). Residual covariance matrix of the first-order discrete-time vector autoregressive (DT-VAR(1)) model.}

\item{Sigma}{Optional (if Stand = 1, then either SigmaVAR, Sigma, or Gamma needed). Residual covariance matrix of the first-order continuous-time (CT-VAR(1)) model, that is, the diffusion matrix.}

\item{Gamma}{Optional (either SigmaVAR, Sigma, or Gamma). Stationary covariance matrix, that is, the contemporaneous covariance matrix of the data.}

\item{Min}{Optional. Minimum time interval used in the Phi-plot. By default, Min = 0.}

\item{Max}{Optional. Maximum time interval used in the Phi-plot. By default, Max = 10.}

\item{Step}{Optional. The step-size taken in the time intervals. By default, Step = 0.05. Hence, using the defaults, the Phi-plots is based on the values of Phi(DeltaT) for DeltaT = 0, 0.05, 0.10, ..., 10. Note: Especially in case of complex eigenvalues, this step size should be very small (then, the oscillating behavior can be seen best).}

\item{WhichElements}{Optional. Matrix of same size as Drift denoting which element/line should be plotted (1) or not (0). By default, WhichElements = NULL. Note that even though not all lines have to be plotted, the full Drift matrix is needed to determine the selected lines.}

\item{Labels}{Optional. Vector with (character) labels of the lines to be plotted. The length of this vector equals the number of 1s in WhichElements (or equals q*q). By default, Labels = NULL, which renders labels with Greek letter of Phi (as a function of the time-interval) together with the indices (of outcome and predictor variables).}

\item{Col}{Optional. Vector with color values (integers) of the lines to be plotted. The length of this vector equals the number of 1s in WhichElements (or equals q*q). By default, Col = NULL, which renders the same color for effects that belong to the same outcome variable (i.e. a row in the Drift matrix). See \url{https://www.statmethods.net/advgraphs/parameters.html} for more information about the values.}

\item{Lty}{Optional. Vector with line type values (integers) of the lines to be plotted. The length of this vector equals the number of 1s in WhichElements (or equals q*q). By default, Lty = NULL, which renders solid lines for the autoregressive effects and the same type of dashed line for reciprocal effects (i.e., same type for Phi_ij as for Phi_ji). See \url{https://www.statmethods.net/advgraphs/parameters.html} for more information about the values.}

\item{Title}{Optional. A character or a list consisting of maximum 3 character-strings or 'expression' class objects that together represent the title of the Phi-plot. By default, Title = NULL, then the following code will be used for the title: as.list(expression(Phi(Delta[t])~plot), "How do the lagged parameters vary", "as a function of the time-interval").}

\item{MaxMinPhi}{Optional. An indicator (TRUE/FALSE) whether vertical lines for the optimum (maximum or minimum) should be added to the plot (TRUE) or not (FALSE). These values are obtained by the function MaxDeltaT(). By default, MaxMinPhi = FALSE; hence, by default, no vertical are added.}
}
\value{
This function returns a Phi-plot for a range of time intervals.
}

\examples{

# library(CTmeta)

### Make Phi-plot ###

## Example 1 ##

# Phi(DeltaT)
DeltaT <- 1             # save the observed time-interval (DeltaT)
Phi <- myPhi[1:2,1:2]   # select a portion of myPhi that corresponds to one primary study
# or: Drift
Drift <- myDrift        # OR select the drift matrix

# Example 1.1: unstandardized Phi #
#
# Make plot of Phi
PhiPlot(DeltaT, Phi)                                           # generate Phi-Plot
PhiPlot(DeltaT, Phi, Min = 0, Max = 10, Step = 0.01)           # specify range of x-axis and precision
PhiPlot(DeltaT, Drift = Drift, Min = 0, Max = 10, Step = 0.01) # using Drift instead of Phi


# Example 1.2: standardized Phi #
q <- dim(Phi)[1]                                      # SigmaVAR must be specified for standardization
SigmaVAR <- diag(q)                                   # for ease                          
PhiPlot(DeltaT, Phi, Stand = 1, SigmaVAR = SigmaVAR)  # standardize Phi using the Stand arugment


## Example 2: input from fitted object of class "varest" ##

DeltaT <- 1           

data <- myData
if (!require("vars")) install.packages("vars")
library(vars)
out_VAR <- VAR(data, p = 1)

# Example 2.1: unstandardized Phi #
PhiPlot(DeltaT, out_VAR)

# Example 2.2: standardized Phi #
PhiPlot(DeltaT, out_VAR, Stand = 1)


## Example 3: Change plot options ##
DeltaT <- 1             
Phi <- myPhi[1:2,1:2]    
q <- dim(Phi)[1]
SigmaVAR <- diag(q) 
#
WhichElements <- matrix(1, ncol = q, nrow = q) # Now, all elements are 1
diag(WhichElements) <- 0                       # Now, the autoregressive parameters are excluded by setting the diagonals to 0.
Lab <- c("12", "21")
Labels <- NULL
for(i in 1:length(Lab)){
 e <- bquote(expression(Phi(Delta[t])[.(Lab[i])]))
 Labels <- c(Labels, eval(e))
}
Col <- c(1,2)
Lty <- c(1,2)

# Standardized Phi
PhiPlot(DeltaT = 1, Phi, Stand = 1, SigmaVAR = SigmaVAR, Min = 0, Max = 10, Step = 0.05, WhichElements = WhichElements, Labels = Labels, Col = Col, Lty = Lty, MaxMinPhi=TRUE)

}
